+++ Kernel/System/Main.pm
@@ -222,6 +222,11 @@ sub FilenameCleanUp {
         return;
     }
 
+    # escape if cleanup is not needed
+    if ( $Param{NoFilenameClean} ) {
+        return $Param{Filename};
+    }
+
     my $Type = lc( $Param{Type} || 'local' );
 
     if ( $Type eq 'md5' ) {
@@ -231,13 +236,16 @@ sub FilenameCleanUp {
     else {
 
         # trim whitespace
-        $Param{Filename} =~ s/^\s+|\n|\s+$//g;
+        $Param{Filename} =~ s/^\s+|\r|\n|\s+$//g;
 
         # strip leading dots
         $Param{Filename} =~ s/^\.+//;
 
         # only whitelisted characters allowed in filename for security
         $Param{Filename} =~ s/[^\w\-+.#_]/_/g;
+
+        # Enclosed alphanumerics are kept on older Perl versions, make sure to replace them too.
+        $Param{Filename} =~ s/[\x{2460}-\x{24FF}]/_/g;
 
         # additional cleaup for attachment filename
         if ( $Type eq 'attachment' ) {
@@ -251,15 +259,36 @@ sub FilenameCleanUp {
             $Param{Filename} =~ s/(\x{00C3}\x{009C}|\x{009C})/Ue/g;
             $Param{Filename} =~ s/(\x{00C3}\x{009F}|\x{00DF})/ss/g;
             $Param{Filename} =~ s/-+/-/g;
+        }
 
-            # cut the string if too long
-            if ( length( $Param{Filename} ) > 100 ) {
-                my $Ext = '';
-                if ( $Param{Filename} =~ /^.*(\.(...|....))$/ ) {
-                    $Ext = $1;
-                }
-                $Param{Filename} = substr( $Param{Filename}, 0, 95 ) . $Ext;
+        # separate filename and extension
+        my $FileName = $Param{Filename};
+        my $FileExt  = '';
+        if ( $Param{Filename} =~ /(.*)\.+(.*)$/ ) {
+
+            # allow extensions up to 10 chars to avoid infinite loop below
+            if (length encode('UTF-8', $2) <= 10) {
+                $FileName = $1;
+                $FileExt  = '.' . $2;
             }
+
+        }
+
+        if ( length $FileName ) {
+            my $ModifiedName;
+
+            # remove character by character starting from the end of the filename string
+            # untill we get acceptable 100 byte long filename size including extension
+            CHOPSTRING:
+            while (1) {
+
+                $ModifiedName = $FileName . $FileExt;
+
+                last CHOPSTRING if ( length encode( 'UTF-8', $ModifiedName ) < 100 );
+                chop $FileName;
+
+            }
+            $Param{Filename} = $ModifiedName;
         }
     }