+++ CHANGES.md
@@ -1,4 +1,5 @@
 #6.0.0.beta1 2017-??-??
+ - 2018-12-03 Fixed infinite loop on filename generation.
  - 2018-05-29 Configurable access to sensitive generic agent ops.
  - 2018-05-29 Generic agent notification handling fixes.
  - 2017-11-13 Fixed incorrect postmaster filter matching.